Adds file header variable replacement feature
Adds feature where variables like $CREATEDYEAR$ and $PROJECT$ are
substituted with dynamic values in the file header. The file header is
updated if a matching header is present, even if the variable's values
have changed. Otherwise, a new header is inserted.

The goal was to preserve all original functionality with the previous
file header update logic, and make this an additive change. As such the
logic is langauge agnostic and works for any style comment syntax or any
arbitrary header with any of the substitution variables provided.