Fix inverted calculation
This looks like it is intending to get the relative line offset, however it was subtracting the line position from the start of the line resulting in a negative number.  These should(tm) be inverted so that the start of the line is subtracted from the place on the line.